Move build script files so the modules are together
