+++ scripts/build.mjs
@@ -4,15 +4,16 @@
  * @license MIT
  * @version 0.1.0
  */
-import buildGitHubPages from './gh-pages.mjs'
+import buildGitHubPages from './modules/gh-pages.mjs'
 buildGitHubPages({
     assets: [
         ['node_modules/github-markdown-css/github-markdown.css', 'styles/github-markdown-css/github-markdown.css'],
         ['node_modules/github-markdown-css/license', 'styles/github-markdown-css/license'],
     ],
     markdown: [
-        'README.md',
-        'content/new-repository-configuration.md',
+        'docs/pull-request/open-source.md',
+        'docs/pull-request/personal.md',
+        'docs/new-repository-configuration.md',
         'contributing/development.md',
         'contributing/self-review.md',
         'contributing/types-of-contributions.md',
@@ -20,6 +21,7 @@ buildGitHubPages({
         'CODE_OF_CONDUCT.md',
         'CONTRIBUTING.md',
         'LICENSE',
+        'README.md',
         'SECURITY.md',
     ],
     directory: 'gh-pages',
